+++ SIL.Core.Tests/Extensions/DateTimeExtensionTests.cs
@@ -137,11 +137,11 @@ namespace SIL.Tests.Extensions
 				Is.EqualTo(new DateTime(2012, 02, 29, 12, 30, 45)));
 		}
 
-		[TestCase("9/3/2025 0:00:00", ExpectedResult = "1482-03-09")]
-		[TestCase("14/4/1482", ExpectedResult = "0939-04-14")]
-		[TestCase("9/3/1800 0:01:00", ExpectedResult = "1257-03-09")]
-		[NonParallelizable]
-		public string ParseDateTimePermissivelyWithException_WithThaiBuddhistCalendar_ReturnsDateWithCorrectYear(string input)
+		[TestCase("2025-05-12")]
+		[TestCase("1482-04-14")]
+		[TestCase("2033-01-02")]
+		public void ParseDateTimePermissivelyWithException_GregorianISO8601WithThaiBuddhistCalendar_ReturnsDateWithCorrectYear(
+			string inputGregorian)
 		{
 			var originalCulture = Thread.CurrentThread.CurrentCulture;
 			try
@@ -151,8 +151,166 @@ namespace SIL.Tests.Extensions
 				buddhistCulture.DateTimeFormat.Calendar = new ThaiBuddhistCalendar();
 				Thread.CurrentThread.CurrentCulture = buddhistCulture;
 
-				return input.ParseDateTimePermissivelyWithException()
-					.ToISO8601TimeFormatDateOnlyString();
+				var result = inputGregorian.ParseDateTimePermissivelyWithException();
+				var inputDateOnly = inputGregorian.Split(' ')[0];
+				var inputDateParts = inputDateOnly.Split('-');
+				Assert.That(inputDateParts.Length, Is.EqualTo(3), "Sanity check");
+				// Note that even when the current culture is Thai/Buddhist, the year,
+				// month and day values in the DateTime object are still Gregorian.
+				Assert.That(result.Year.ToString(), Is.EqualTo(inputDateParts[0]));
+				Assert.That(result.Month, Is.EqualTo(int.Parse(inputDateParts[1])));
+				Assert.That(result.Day, Is.EqualTo(int.Parse(inputDateParts[2])));
+				Assert.That(result.ToISO8601TimeFormatDateOnlyString(),
+					Is.EqualTo(inputDateOnly));
+
+				var expectedResultFormattedAsBuddhistDate =
+					$"{result.Day}/{result.Month}/{result.Year + 543}";
+				Assert.That(result.ToShortDateString(),
+					Is.EqualTo(expectedResultFormattedAsBuddhistDate));
+			}
+			finally
+			{
+				// Reset system locale to the original culture
+				Thread.CurrentThread.CurrentCulture = originalCulture;
+			}
+		}
+
+		/// <summary>
+		/// Input dates are not formatted as valid ISO8601 dates. Method should guess which
+		/// calendar to use based on year and current culture. If year is in the "modern" range
+		/// for the Gregorian calendar, it should be treated as a Gregorian date. If year is
+		/// older or way in the future, it should be treated as a Buddhist date. Thai culture
+		/// normally formats dates as dd/MM/yyyy, so that is how these dates should be
+		/// interpreted.
+		/// </summary>
+		/// <param name="input">The ambiguous input date, not formatted as ISO 8601</param>
+		/// <param name="expectedToInterpretAsGregorianYear">Flag indicating whether algorithm is
+		/// expected to guess that the input date represents a year in the Gregorian calendar
+		/// </param>
+		/// <param name="expectedResultFormattedAsBuddhistDate"></param>
+		[TestCase("9/3/2025 0:00:00", true, "9/3/2568")]
+		[TestCase("2-2-1920", true, "2/2/2463")] // 
+		[TestCase("14/4/1543", false, "14/4/1543")]
+		[TestCase("3/9/1257 0:01:00", false, "3/9/1257")]
+		[TestCase("9/3/2568 0:01:00", false, "9/3/2568")]
+		[NonParallelizable]
+		public void ParseDateTimePermissivelyWithException_WithThaiBuddhistCalendar_ReturnsDateWithCorrectYear(
+			string input, bool expectedToInterpretAsGregorianYear,
+			string expectedResultFormattedAsBuddhistDate)
+		{
+			var originalCulture = Thread.CurrentThread.CurrentCulture;
+			try
+			{
+				// Temporarily set system locale to use Buddhist date system
+				var buddhistCulture = new CultureInfo("th-TH");
+				buddhistCulture.DateTimeFormat.Calendar = new ThaiBuddhistCalendar();
+				Thread.CurrentThread.CurrentCulture = buddhistCulture;
+
+				var result = input.ParseDateTimePermissivelyWithException();
+				Assert.That(result.ToShortDateString(),
+					Is.EqualTo(expectedResultFormattedAsBuddhistDate));
+				var inputDateOnly = input.Split(' ')[0];
+				var inputDateParts = inputDateOnly.Split('/', '-');
+				Assert.That(inputDateParts.Length, Is.EqualTo(3), "Sanity check");
+				// Note that even when the current culture is Thai/Buddhist, the unformatted year,
+				// month and day stored in the DateTime object are still Gregorian.
+				var expectedGregorianYear = expectedToInterpretAsGregorianYear
+					? inputDateParts[2]
+					: (int.Parse(inputDateParts[2]) - 543).ToString();
+				Assert.That(result.Year.ToString(), Is.EqualTo(expectedGregorianYear));
+				Assert.That(result.Month.ToString(), Is.EqualTo(inputDateParts[1]));
+				Assert.That(result.Day.ToString(), Is.EqualTo(inputDateParts[0]));
+				Assert.That(result.ToISO8601TimeFormatDateOnlyString(), Is.EqualTo(
+					$"{int.Parse(expectedGregorianYear):D4}-{int.Parse(inputDateParts[1]):D2}-{int.Parse(inputDateParts[0]):D2}"));
+			}
+			finally
+			{
+				// Reset system locale to the original culture
+				Thread.CurrentThread.CurrentCulture = originalCulture;
+			}
+		}
+
+		/// <summary>
+		/// For these test cases, well will generate a future date (beyond the
+		/// numberOfDaysIntoFutureReasonablyExpected) and pass a value of
+		/// oldestReasonablyExpectedYear that is more than 543 years ago. Thus, we
+		/// expect the algorithm to guess that these are (ancient) Buddhist dates as
+		/// opposed to future Gregorian dates.
+		/// </summary>
+		/// <remarks>Given the non-Thai (probably usually English) month name when using the
+		/// "dd MMM yyyy" format, it's slightly surprising that the Thai/Buddhist locale can parse
+		/// it, but apparently it can.</remarks>
+		[TestCase("dd MMM yyyy")] // Medium pattern (e.g. 14 May 2025)
+		[TestCase("dd/MM/yyyy")] // Thai/European-style numeric date, zero-padded (e.g. 14/05/2025)
+		[TestCase("d/M/yyyy")] // Thai/European-style numeric date (e.g. 14/5/2025)
+		[TestCase("d-M-yyyy")] // Thai/European-style numeric date with dashes (e.g. 14-5-2025)
+		[NonParallelizable]
+		public void ParseDateTimePermissivelyWithException_NearFutureDatesWithThaiBuddhistCalendar_ReturnsDateWithPastYear(string inputFormat)
+		{
+			var futureDateA = DateTime.Today.AddDays(2);
+			var futureDateB = DateTime.Today.AddDays(5);
+			string inputA = futureDateA.ToString(inputFormat);
+			string inputB = futureDateB.ToString(inputFormat);
+			int expectedYearA = futureDateA.Year - 543;
+			int expectedYearB = futureDateB.Year - 543;
+			var originalCulture = Thread.CurrentThread.CurrentCulture;
+			try
+			{
+				// Temporarily set system locale to use Buddhist date system
+				var buddhistCulture = new CultureInfo("th-TH");
+				buddhistCulture.DateTimeFormat.Calendar = new ThaiBuddhistCalendar();
+				Thread.CurrentThread.CurrentCulture = buddhistCulture;
+
+				var result = inputA.ParseDateTimePermissivelyWithException(
+					oldestReasonablyExpectedYear: 1481);
+				Assert.That(result.Year, Is.EqualTo(expectedYearA));
+				result = inputB.ParseDateTimePermissivelyWithException(
+					oldestReasonablyExpectedYear: 1481,
+					numberOfDaysIntoFutureReasonablyExpected: 4);
+				Assert.That(result.Year, Is.EqualTo(expectedYearB));
+			}
+			finally
+			{
+				// Reset system locale to the original culture
+				Thread.CurrentThread.CurrentCulture = originalCulture;
+			}
+		}
+
+		/// <summary>
+		/// For these test cases, well will generate a future date (beyond the
+		/// numberOfDaysIntoFutureReasonablyExpected). However, these are date formats that cannot
+		/// be parsed in Thai/Buddhist culture, so we expect them to be interpreted as Gregorian
+		/// dates.
+		/// </summary>
+		[TestCase("d")] // Short date pattern (e.g. 14/5/2025)
+		[TestCase("D")] // Long date pattern (e.g. Wednesday, 14 May 2025)
+		[TestCase("dddd, dd MMMM yyyy")] // Full long format (e.g. Wednesday, 14 May 2025)
+		[TestCase("M/d/yyyy")] // US-style numeric date (e.g. 5/14/2025)
+		[TestCase("M-d-yyyy")] // US-style numeric date with dashes (e.g. 5-14-2025)
+		[TestCase("MM/dd/yyyy")] // US-style numeric date, zero-padded (e.g. 05/14/2025)
+		[TestCase("MM-dd-yyyy")] // US-style numeric date with dashes, zero-padded (e.g. 05-14-2025)
+		[NonParallelizable]
+		public void ParseDateTimePermissivelyWithException_NearFutureUSDatesWithThaiBuddhistCalendar_ReturnsDateWithPastYear(string inputFormat)
+		{
+			var futureDateA = DateTime.Today.AddDays(2);
+			var futureDateB = DateTime.Today.AddDays(5);
+			string inputA = futureDateA.ToString(inputFormat);
+			string inputB = futureDateB.ToString(inputFormat);
+			int expectedYearA = futureDateA.Year;
+			int expectedYearB = futureDateB.Year;
+			var originalCulture = Thread.CurrentThread.CurrentCulture;
+			try
+			{
+				// Temporarily set system locale to use Buddhist date system
+				var buddhistCulture = new CultureInfo("th-TH");
+				buddhistCulture.DateTimeFormat.Calendar = new ThaiBuddhistCalendar();
+				Thread.CurrentThread.CurrentCulture = buddhistCulture;
+
+				var result = inputA.ParseDateTimePermissivelyWithException();
+				Assert.That(result.Year, Is.EqualTo(expectedYearA));
+				result = inputB.ParseDateTimePermissivelyWithException(
+					numberOfDaysIntoFutureReasonablyExpected: 4);
+				Assert.That(result.Year, Is.EqualTo(expectedYearB));
 			}
 			finally
 			{
@@ -179,7 +337,7 @@ namespace SIL.Tests.Extensions
 		[TestCase("@13:00.T")]
 		public void ParseDateTimePermissivelyWithException_UnknownFormat_ThrowsApplicationException(string input)
 		{
-			Assert.That(input.ParseDateTimePermissivelyWithException,
+			Assert.That(() => input.ParseDateTimePermissivelyWithException(),
 				Throws.TypeOf<ApplicationException>().With.InnerException.TypeOf<FormatException>());
 		}
 	}